+++ src/app/shared/NoteContentKind32123.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  import {annotateMedia} from "src/util/misc"
   import Card from "src/partials/Card.svelte"
   import Media from "src/partials/Media.svelte"
 
@@ -10,5 +9,5 @@
 
 <Card interactive class="flex flex-col gap-4" on:click={() => window.open(link)}>
   <p class="staatliches text-xl">{title} | {creator}</p>
-  <Media link={annotateMedia(enclosure)} />
+  <Media url={enclosure} />
 </Card>