+++ src/app/shared/NoteImages.svelte
@@ -1,0 +1,31 @@
+<script lang="ts">
+  import {without} from "ramda"
+  import Media from "src/partials/Media.svelte"
+
+  export let compose
+  export let value = []
+
+  const getUrl = imeta => imeta.type("url").values().first()
+
+  export const addImage = imeta => {
+    value = value.concat(imeta)
+    compose.write("\n" + getUrl(imeta))
+  }
+
+  export const removeImage = imeta => {
+    const content = compose.parse()
+
+    compose.clear()
+    compose.write(content.replace(getUrl(imeta), ""))
+
+    value = without([imeta], value)
+  }
+</script>
+
+<div class="columns-2 gap-2 lg:columns-3">
+  {#each value as imeta}
+    <div class="pb-2">
+      <Media {imeta} url={getUrl(imeta)} onClose={() => removeImage(imeta)} />
+    </div>
+  {/each}
+</div>