+++ src/app/shared/NoteReply.svelte
@@ -1,14 +1,14 @@
 <script lang="ts">
   import {Tags} from "paravel"
   import {createEventDispatcher} from "svelte"
-  import {without, identity, uniq} from "ramda"
+  import {join, without, identity, uniq} from "ramda"
   import {getGroupAddress, asNostrEvent} from "src/util/nostr"
   import {slide} from "src/util/transition"
   import ImageInput from "src/partials/ImageInput.svelte"
   import Chip from "src/partials/Chip.svelte"
-  import Media from "src/partials/Media.svelte"
   import Compose from "src/app/shared/Compose.svelte"
   import NoteOptions from "src/app/shared/NoteOptions.svelte"
+  import NoteImages from "src/app/shared/NoteImages.svelte"
   import {
     Publisher,
     buildReply,
@@ -25,11 +25,10 @@
 
   const dispatch = createEventDispatcher()
 
-  let data = null
-  let reply = null
-  let container = null
+  let images, compose, container, options
+  let isOpen = false
+  let mentions = []
   let draft = ""
-  let options
   let opts = {
     warning: "",
     groups: parent.wrap ? Tags.from(parent).communities().all() : [],
@@ -41,15 +40,13 @@
   export const start = () => {
     dispatch("start")
 
-    data = {
-      image: null,
-      mentions: without(
-        [$session.pubkey],
-        uniq(Tags.from(parent).type("p").values().all().concat(parent.pubkey))
-      ),
-    }
+    isOpen = true
+    mentions = without(
+      [$session.pubkey],
+      uniq(Tags.from(parent).type("p").values().all().concat(parent.pubkey))
+    )
 
-    setTimeout(() => reply.write(draft))
+    setTimeout(() => compose.write(draft))
   }
 
   const setOpts = e => {
@@ -57,8 +54,8 @@
   }
 
   const saveDraft = () => {
-    if (reply) {
-      draft = reply.parse()
+    if (compose) {
+      draft = compose.parse()
     }
   }
 
@@ -69,24 +66,26 @@
   const reset = () => {
     dispatch("reset")
 
-    data = null
-    reply = null
+    mentions = null
+    compose = null
   }
 
   const removeMention = pubkey => {
-    data.mentions = without([pubkey], data.mentions)
+    mentions = without([pubkey], mentions)
   }
 
-  const getContent = () => (reply.parse() + "\n" + (data.image || "")).trim()
-
   const send = async () => {
-    const content = getContent()
+    const content = compose.parse().trim()
 
     if (!content) {
       return
     }
 
-    const tags = data.mentions.map(mention)
+    const tags = mentions.map(mention)
+
+    for (const imeta of images.value) {
+      tags.push(["imeta", ...imeta.all().map(join(" "))])
+    }
 
     if (opts.warning) {
       tags.push(["content-warning", opts.warning])
@@ -120,7 +119,7 @@
 
 <svelte:body on:click={onBodyClick} />
 
-{#if data}
+{#if isOpen}
   <div
     transition:slide|local
     class="note-reply relative z-10 my-2 flex flex-col gap-1"
@@ -130,8 +129,8 @@
       <div class="absolute bottom-0 left-4 top-0 z-0 -my-2 w-px bg-gray-6" />
     {/if}
     <div class="z-10 overflow-hidden rounded-2xl border border-solid border-gray-6">
-      <div class="bg-gray-7 p-3 text-gray-2" class:rounded-b={data.mentions.length === 0}>
-        <Compose bind:this={reply} onSubmit={send} style="min-height: 4rem">
+      <div class="bg-gray-7 p-3 text-gray-2" class:rounded-b={mentions.length === 0}>
+        <Compose bind:this={compose} onSubmit={send} style="min-height: 4rem">
           <div class="flex flex-col justify-start" slot="addon">
             <button
               on:click={send}
@@ -141,20 +140,14 @@
           </div>
         </Compose>
       </div>
-      {#if data.image}
-        <div class="bg-gray-7 p-2">
-          <Media
-            link={{type: "image", url: data.image}}
-            onClose={() => {
-              data.image = null
-            }} />
-        </div>
-      {/if}
+      <div class="bg-gray-7 p-2">
+        <NoteImages bind:this={images} bind:compose />
+      </div>
       <div class="h-px bg-gray-7 group-[.modal]:bg-gray-6" />
       <div class="flex gap-2 rounded-b bg-gray-7 p-2 text-sm text-gray-2">
         <div class="inline-block border-r border-solid border-gray-6 py-2 pl-1 pr-3">
           <div class="flex cursor-pointer items-center gap-3">
-            <ImageInput bind:value={data.image}>
+            <ImageInput multi hostLimit={3} on:change={e => images.addImage(e.detail)}>
               <i slot="button" class="fa fa-paperclip" />
             </ImageInput>
             <i class="fa fa-cog" on:click={() => options.setView("settings")} />
@@ -162,7 +155,7 @@
           </div>
         </div>
         <div on:click|stopPropagation>
-          {#each data.mentions as pubkey}
+          {#each mentions as pubkey}
             <Chip class="mb-1 mr-1" theme="dark" onRemove={() => removeMention(pubkey)}>
               {displayPubkey(pubkey)}
             </Chip>