+++ src/engine/media/commands.ts
@@ -1,27 +1,86 @@
-import {Fetch} from "hurdak"
-import {createEvent} from "paravel"
-import {generatePrivateKey} from "nostr-tools"
-import {signer} from "src/engine/session/derived"
+import {prop, nth, uniq, flatten, groupBy} from "ramda"
+import {Fetch, tryFunc, sleep} from "hurdak"
+import {now, cached, Tags} from "paravel"
+import {joinPath} from "src/util/misc"
+import type {Event} from "src/engine/events/model"
+import {nip98Fetch} from "src/engine/auth/commands"
+import {stripExifData, blobToFile} from "src/util/html"
 
-export const uploadToNostrBuild = async body => {
-  const $signer = signer.get()
-  const url = "https://nostr.build/api/v2/upload/files"
-  const template = createEvent(27235, {
-    tags: [
-      ["u", url],
-      ["method", "POST"],
-    ],
+export const getMediaProviderURL = cached({
+  maxSize: 10,
+  getKey: ([url]) => url,
+  getValue: ([url]) => fetchMediaProviderURL(url),
+})
+
+const fetchMediaProviderURL = async host =>
+  prop("api_url", await Fetch.fetchJson(joinPath(host, ".well-known/nostr/nip96.json")))
+
+const fileToFormData = file => {
+  const formData = new FormData()
+
+  formData.append("file[]", file)
+
+  return formData
+}
+
+export const uploadFileToHost = async (url, file) => {
+  const startTime = now()
+  const apiUrl = await getMediaProviderURL(url)
+  const response = await nip98Fetch(apiUrl, "POST", fileToFormData(file))
+
+  // If the media provider uses delayed processing, we need to wait for the processing to be done
+  while (response.processing_url) {
+    const {status, nip94_event} = await nip98Fetch(response.processing_url, "GET")
+
+    if (status === "success") {
+      return Tags.from(nip94_event).type("url").values().first()
+    }
+
+    if (now() - startTime > 60) {
+      break
+    }
+
+    await sleep(3000)
+  }
+
+  return response.nip94_event
+}
+
+export const uploadFilesToHost = (url, files) =>
+  Promise.all(files.map(file => tryFunc(async () => await uploadFileToHost(url, file))))
+
+export const uploadFileToHosts = (urls, file) =>
+  Promise.all(urls.map(url => tryFunc(async () => await uploadFileToHost(url, file))))
+
+export const uploadFilesToHosts = async (urls, files) =>
+  flatten(await Promise.all(urls.map(url => uploadFilesToHost(url, files))))
+
+export const compressFiles = (files, opts) =>
+  Promise.all(
+    files.map(async f => {
+      if (f.type.match("image/(webp|gif)")) {
+        return f
+      }
+
+      return blobToFile(await stripExifData(f, opts))
+    })
+  )
+
+export const eventsToMeta = (events: Event[]) => {
+  const tagsByHash = groupBy(
+    tags => tags.type("ox").values().first(),
+    events.map(e => Tags.from(e))
+  )
+
+  // Merge all nip94 tags together so we can supply as much imeta as possible
+  return Object.values(tagsByHash).map(groupedTags => {
+    return new Tags(uniq(groupedTags.flatMap(tags => tags.filter(nth(1)).all())))
   })
+}
 
-  const event = await ($signer.canSign()
-    ? $signer.signAsUser(template)
-    : $signer.signWithKey(template, generatePrivateKey()))
+export const uploadFiles = async (urls, files, compressorOpts = {}) => {
+  const compressedFiles = await compressFiles(files, compressorOpts)
+  const nip94Events = await uploadFilesToHosts(urls, compressedFiles)
 
-  return Fetch.fetchJson(url, {
-    body,
-    method: "POST",
-    headers: {
-      Authorization: `Nostr ${btoa(JSON.stringify(event))}`,
-    },
-  })
+  return eventsToMeta(nip94Events)
 }