+++ src/engine/session/utils/settings.ts
@@ -11,6 +11,7 @@ export const getDefaultSettings = () => ({
   auto_authenticate: false,
   min_wot_score: 1,
   enable_reactions: true,
+  nip96_urls: env.get().NIP96_URLS,
   imgproxy_url: env.get().IMGPROXY_URL,
   dufflepud_url: env.get().DUFFLEPUD_URL,
   multiplextr_url: env.get().MULTIPLEXTR_URL,