+++ src/partials/Media.svelte
@@ -1,19 +1,23 @@
 <script lang="ts">
+  import cx from "classnames"
   import {last} from "ramda"
   import {ellipsize, Fetch} from "hurdak"
+  import {Tags} from "paravel"
   import {AudioController} from "src/util/audio"
   import Audio from "src/partials/Audio.svelte"
+  import Image from "src/partials/Image.svelte"
   import Anchor from "src/partials/Anchor.svelte"
   import Spinner from "src/partials/Spinner.svelte"
   import {dufflepud, imgproxy} from "src/engine"
 
-  export let link
+  export let url
+  export let imeta = new Tags([["url", url]])
   export let onClick = null
   export let onClose = null
   export let fullSize = false
 
   const loadPreview = async () => {
-    const json = await Fetch.postJson(dufflepud("link/preview"), {url: link.url})
+    const json = await Fetch.postJson(dufflepud("link/preview"), {url})
 
     if (!json.title && !json.image) {
       throw new Error("Unable to load preview")
@@ -23,24 +27,18 @@
   }
 </script>
 
-{#if link.type === "audio"}
-  <Audio controller={new AudioController(link.url)} />
+{#if url.match(/\.(wav|mp3|m3u8)$/)}
+  <Audio controller={new AudioController(url)} />
 {:else}
   <Anchor
     external
     type="unstyled"
-    href={onClick ? null : link.url}
+    href={onClick ? null : url}
     on:click={onClick}
     style="background-color: rgba(15, 15, 14, 0.5)"
     class="relative flex flex-col overflow-hidden rounded-xl border border-solid border-gray-6">
-    {#if link.type === "image"}
-      <img
-        alt="Link preview"
-        src={imgproxy(link.url)}
-        class:max-h-96={!fullSize}
-        class="object-contain object-center" />
-    {:else if link.type === "spotify"}
-      {@const id = last(link.url.split("?")[0].match(/[a-z]+\/[0-9A-z]+$/))}
+    {#if url.match(/open.spotify.com/)}
+      {@const id = last(url.split("?")[0].match(/[a-z]+\/[0-9A-z]+$/))}
       {@const src = `https://open.spotify.com/embed/${id}`}
       <iframe
         {src}
@@ -51,8 +49,13 @@
         frameBorder="0"
         allow="autoplay; clipboard-write; encrypted-media; fullscreen; picture-in-picture"
         loading="lazy" />
-    {:else if link.type === "video"}
-      <video controls src={link.url} class="max-h-96 object-contain object-center" />
+    {:else if url.match(/\.(mov|webm|mp4)$/)}
+      <video controls src={url} class="max-h-96 object-contain object-center" />
+    {:else if url.match(/\.(jpe?g|png|gif|webp)$/)}
+      <Image
+        alt="Link preview"
+        src={imeta.type("url").values().all().map(imgproxy)}
+        class={cx("object-contain object-center", {"max-h-96": !fullSize})} />
     {:else}
       {#await loadPreview()}
         <Spinner />
@@ -72,7 +75,7 @@
         {/if}
       {:catch}
         <p class="mb-1 px-12 py-24 text-center text-gray-5">
-          Unable to load a preview for {link.url}
+          Unable to load a preview for {url}
         </p>
       {/await}
     {/if}